+++ src/Sekiban.Core/Command/ICommandWithStaticHandlerCommon.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Core.Command;
+
+public interface ICommandWithStaticHandlerCommon;
